-pool of stages -extra properties tests - build. -flag --noVerify to skip the tests -unit-tests
