--- lib/context/context.go
@@ -1,12 +1,0 @@
-package context
-
-import (
-	kingpin "gopkg.in/alecthomas/kingpin.v2"
-
-	"github.com/Travix-International/Travix.Core.Adk/lib/config"
-)
-
-type Context struct {
-	App    *kingpin.Application
-	Config *config.Config
-}
